app/requirements.txt
streamlit==1.31.0
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2
PyMuPDF==1.23.8
python-docx==1.1.0
scikit-learn==1.4.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz


